+++ protoc_plugin/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 21.0.0
+
+* Identifiers `fromBuffer`, `fromJson`, `$_defaultFor`, `initByValue` are no
+  longer reserved. Proto fields with those Dart names will no longer have a
+  suffix added. [#xxx]
+
 ## 20.0.1
 
 * Fix proto3 repeated field encoding without the `packed` option ([#345],